Refactor the arguments to combine them automatically
